+++ FMWKubernetesMAA/OracleEnterpriseDeploymentAutomation/OracleIdentityManagement/utils/delete_oam.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-# Copyright (c) 2021, 2022, Oracle and/or its affiliates.
+# Copyright (c) 2021, 2023, Oracle and/or its affiliates.
 # Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 #
 # This is an example of a script which will delete an OAM deployment
@@ -7,13 +7,48 @@
 # Dependencies: ../common/functions.sh
 #               ../responsefile/idm.rsp
 #
-# Usage: delete_oam.sh
+# Usage: delete_oam.sh [-r responsefile -p passwordfile]
 #
 
-MYDIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )" 
+SCRIPTDIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )" 
+SCRIPTDIR=$SCRIPTDIR/..
 
-. $MYDIR/../common/functions.sh
+while getopts 'r:p:' OPTION
+do
+  case "$OPTION" in
+    r)
+      RSPFILE=$SCRIPTDIR/responsefile/$OPTARG
+     ;;
+    p)
+      PWDFILE=$SCRIPTDIR/responsefile/$OPTARG
+     ;;
+    ?)
+     echo "script usage: $(basename $0) [-r responsefile -p passwordfile] " >&2
+     exit 1
+     ;;
+   esac
+done
+
+
+RSPFILE=${RSPFILE=$SCRIPTDIR/responsefile/idm.rsp}
+PWDFILE=${PWDFILE=$SCRIPTDIR/responsefile/.idmpwds}
+
 . $RSPFILE
+if [ $? -gt 0 ]
+then
+    echo "Responsefile : $RSPFILE does not exist."
+    exit 1
+fi
+
+. $PWDFILE
+if [ $? -gt 0 ]
+then
+    echo "Passwordfile : $PWDFILE does not exist."
+    exit 1
+fi
+
+. $SCRIPTDIR/common/functions.sh
+
 WORKDIR=$LOCAL_WORKDIR/OAM
 LOGDIR=$WORKDIR/logs
 
@@ -95,23 +130,24 @@ else
 fi
 
 
-print_time STEP "Drop Schemas" $ST $ET 
-
 # Delete All contents in the Persistent Volumes
 # Requires that the PV is mounted locally
 
-ST=`date +%s`
 echo "Deleting Volumes"
-echo " Deleting  $OAM_LOCAL_SHARE/*">> $LOG 2>&1
-rm -rf $OAM_LOCAL_SHARE/>> $LOG 2>&1
-echo " Deleting  $WORKDIR/*">> $LOG 2>&1
-rm  -rf $WORKDIR/* 
-echo " Deleting  $LOCAL_WORKDIR/OHS ">> $LOG 2>&1
-rm  -rf $LOCAL_WORKDIR/OHS/*/login_vh.conf $LOCAL_WORKDIR/OHS/*/iadadmin_vh.conf $LOCAL_WORKDIR/oam_installed>> $LOG 2>&1
-echo " Deleting  Complete">> $LOG 2>&1
 
-ET=`date +%s`
-print_time STEP "Delete Volume" $ST $ET 
+if [ ! "$OAM_LOCAL_SHARE" = "" ]
+then
+  rm -rf $OAM_LOCAL_SHARE/>> $LOG 2>&1
+fi
+
+
+if [ ! "$WORKDIR" = "" ] && [ ! "$LOGDIR" = "" ] && [ ! "$LOCAL_WORKDIR" = "" ]
+then
+  rm  -rf $WORKDIR/* 
+  rm  -rf $LOCAL_WORKDIR/OHS/*/login_vh.conf $LOCAL_WORKDIR/OHS/*/iadadmin_vh.conf $LOCAL_WORKDIR/oam_installed>> $LOG 2>&1
+else
+  echo "Unable to delete volumes."
+fi
 
 # Remove Persistent Volume & Claim from Kubernetes
 #
