+++ FMWKubernetesMAA/OracleEnterpriseDeploymentAutomation/OracleIdentityManagement/utils/delete_ohs.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-# Copyright (c) 2022, Oracle and/or its affiliates.
+# Copyright (c) 2022, 2023, Oracle and/or its affiliates.
 # Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 #
 # This is an example of a script which will delete an OHS deployment
@@ -8,16 +8,50 @@
 #               ../common/ohs_functions.sh
 #               ../responsefile/idm.rsp
 #
-# Usage: delete_ohs.sh
+# Usage: delete_ohs.sh  [-r responsefile -p passwordfile]
 #
-MYDIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+SCRIPTDIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+SCRIPTDIR=$SCRIPTDIR/..
 
-. $MYDIR/../common/functions.sh
-. $MYDIR/../common/ohs_functions.sh
+while getopts 'r:p:' OPTION
+do
+  case "$OPTION" in
+    r)
+      RSPFILE=$SCRIPTDIR/responsefile/$OPTARG
+     ;;
+    p)
+      PWDFILE=$SCRIPTDIR/responsefile/$OPTARG
+     ;;
+    ?)
+     echo "script usage: $(basename $0) [-r responsefile -p passwordfile] " >&2
+     exit 1
+     ;;
+   esac
+done
+
+
+RSPFILE=${RSPFILE=$SCRIPTDIR/responsefile/idm.rsp}
+PWDFILE=${PWDFILE=$SCRIPTDIR/responsefile/.idmpwds}
+
 . $RSPFILE
-TEMPLATE_DIR=$SCRIPTDIR/templates/ohs
-WORKDIR=$LOCAL_WORKDIR/OHS
+if [ $? -gt 0 ]
+then
+    echo "Responsefile : $RSPFILE does not exist."
+    exit 1
+fi
 
+. $PWDFILE
+if [ $? -gt 0 ]
+then
+    echo "Passwordfile : $PWDFILE does not exist."
+    exit 1
+fi
+
+. $SCRIPTDIR/common/functions.sh
+. $SCRIPTDIR/common/ohs_functions.sh
+
+WORKDIR=$LOCAL_WORKDIR/OHS
+TEMPLATE_DIR=$SCRIPTDIR/templates/ohs
 
 mkdir $LOCAL_WORKDIR/deleteLogs > /dev/null 2>&1
 
@@ -51,13 +85,13 @@ then
     delete_instance $OHS_HOST1 $OHS1_NAME >> $LOG 2>&1
 
     echo "Deleting OHS Instance Files on $OHS1_NAME"
-    ssh $OHS_HOST1 "rm -rf $OHS_DOMAIN" >> $LOG 2>&1
+    $SSH ${OHS_USER}@$OHS_HOST1 "rm -rf $OHS_DOMAIN" >> $LOG 2>&1
 
     echo "Deleting OHS Install on $OHS_HOST1"
-    ssh $OHS_HOST1 "rm -rf $OHS_ORACLE_HOME" >> $LOG 2>&1
+    $SSH ${OHS_USER}@$OHS_HOST1 "rm -rf $OHS_ORACLE_HOME" >> $LOG 2>&1
 
     echo "Deleting Oracle Inventory on $OHS_HOST1"
-    ssh $OHS_HOST1 "rm -rf $OHS_BASE/oraInventory" >> $LOG 2>&1
+    $SSH ${OHS_USER}@$OHS_HOST1 "rm -rf $OHS_BASE/oraInventory" >> $LOG 2>&1
 fi
 
 if [ ! "$OHS_HOST2" = "" ]
@@ -71,17 +105,22 @@ then
     delete_instance $OHS_HOST2 $OHS2_NAME >> $LOG 2>&1
 
     echo "Deleting OHS Instance files on $OHS2_NAME"
-    ssh $OHS_HOST2 "rm -rf $OHS_DOMAIN" >> $LOG 2>&1
+    $SSH ${OHS_USER}@$OHS_HOST2 "rm -rf $OHS_DOMAIN" >> $LOG 2>&1
 
     echo "Deleting OHS Install on $OHS_HOST2"
-    ssh $OHS_HOST2 "rm -rf $OHS_ORACLE_HOME" >> $LOG 2>&1
+    $SSH ${OHS_USER}@$OHS_HOST2 "rm -rf $OHS_ORACLE_HOME" >> $LOG 2>&1
 
     echo "Deleting Oracle Inventory on $OHS_HOST2"
-    ssh $OHS_HOST2 "rm -rf $OHS_BASE/oraInventory" >> $LOG 2>&1
+    $SSH ${OHS_USER}@$OHS_HOST2 "rm -rf $OHS_BASE/oraInventory" >> $LOG 2>&1
 fi
 
 echo "Delete Working Directory"
-rm -rf $LOCAL_WORKDIR/OHS $LOCAL_WORKDIR/ohs_installed
+if  [ ! "$LOCAL_WORKDIR" = "" ]
+then
+  rm -rf $LOCAL_WORKDIR/OHS $LOCAL_WORKDIR/ohs_installed
+else
+  echo "Unable to Delete Working Directory."
+fi
 
 
 FINISH_TIME=`date +%s`
